Issue number:  40
- Changes to COM4J:
  - Added a new mechanism to add shutdown tasks. If you need to release resources upon shutdown of the application then add a Com4J shutdown task rather than a Runtime shutdown hook. The reason:
    If you add your shutdown task via Runtime.getRuntime().addShutdownHook(..) it is likely that the shutdown hooks of the Com4J runtime is running before your own shutdown hook. This means your shutdown hook cannot access any Com4J resources anymore...
	For Com4J developers: There are two kinds of shutdown tasks:
	 - application shutdown tasks that are added by the users of Com4J and
	 - com4j shutdown tasks that shut down the Com4J runtime.
- Added a class LiveObjectCollection to collect all living objects of a thread.
- Com4JObject: added getPtr(). This way COM4J.unwrap(..) is unnecessary.
- Changes to ComThread:
  - BugFix: issue 40, tasks are now executed FIFO instead of LIFO
  - BugFix: if the task was interrupted, then the result is invalid
  - BugFix: Adding a Com4JShutdownTask instead of a JVM shutdown hook and disposing all objects instead of just killing all threads.
  - Remember the living objects by reference in an LiveObjectCollection (instead of just counting living objects)
  - kill() is now killing the ComThread by setting "die" to true directly
  - access to the field "threads" are synchronized
- Win32Lock: added a version with a timeout parameter
- Changes to Wrapper:
  - BugFix: dispose was calling Native.release even if the pointer was NULL.
- BugFix: ComCollection (IEnumVARIANT): We need to remember the marshaling thread to be able to interpret  returned Variant as an interface pointer to a COM object
- lots of JavaDoc


git-svn-id: https://svn.java.net/svn/com4j~svn/trunk/com4j@349 590aa308-3dd6-4388-9d27-cebafe3e2bdc

+++ tlbimp/src/com4j/tlbimp/FileCodeWriter.java
@@ -16,7 +16,12 @@ public class FileCodeWriter implements CodeWriter {
 
     public IndentingWriter create(File file) throws IOException {
         file = new File(outDir,file.getPath());
-        file.getParentFile().mkdirs();
+        File dir = file.getParentFile();
+        boolean newCreated = dir.mkdirs();
+        boolean exists = dir.exists();
+        if(!exists){
+          throw new IOException("Could not create the directory "+ file.getParentFile().getAbsolutePath());
+        }
         // TODO: proper escaping
         return new IndentingWriter(new FileWriter(file));
     }
